goto文を実装した

+++ source/smilebasic/vm/types.d
@@ -7,5 +7,5 @@ struct Pointer{
   ///スロット
   int slot;
   ///プログラムカウンタ
-  ulong count; 
+  uint count; 
 }
